free系の関数宣言をfree.hに集約
