+++ src/token_utils.c
@@ -32,19 +32,3 @@ bool	is_redir_token(t_token *token)
 	return (false);
 }
 
-void	free_tokens(t_token *token)
-{
-	t_token	*next;
-
-	while (token)
-	{
-		next = token->next;
-		if (is_word_or_quoted_token(token) || is_io_num_token(token))
-		{
-			free(token->word);
-			token->word = NULL;
-		}
-		free(token);
-		token = next;
-	}
-}
